extract LastPurchaseViewModel to separate file
